requirements.txt
pytest==8.3.2
websockets==13.0.1
pycryptodome==3.20.0
asyncio==3.4.3
pytest-asyncio==0.24.0
inquirer